Replacing urls which depend on the context of the target object with it's proper url.
This fixes the issue with broken images in news items and seems to
treat links correctly etc. I tried to parse the view, extract just the
elements of interest and convert the result to html
again. lxml.etree.fromstring complains about invalid entities
e.g. mdash, and parsing with the lxml soupparser and then converting
back to html with lxml.etree.tostring breaks the layout
significantly.