Revert the layout so that the unsubscribe form is separate and at the bottom of the page.
FTR This was explicitly requested by OSHA despite the fact that the link to the unsubscribe form typically does not do anything. The unsubscribe form appears at the end of the page, so it would only be scrolled into view (at the bottom of the page) if the list of "older newsletters" has been expanded, or if the height of the visitor's browser window is approx. 600px.