Refs #9323 OSHmail use context.translate for translation
The language to be used for translating the content is determined by
the context, but we store oshmails in /data so the fallback language
is used (en). To work around this I've used context.translate. Since
the strings cannot be extracted without the i18n:translate attr, I've
added the extra strings to osha.policy.missing_i18n_labels.py